Use a user-sepcified Java VM instead of the system default Java VM, and pass this on in narwhal.conf, so that narwhal also uses the bootstrapped Java VM.
Users may either set JAVA_HOME in the environment
or pass --java-home as a parameter to specify
where the Java VM resides.

check_build_environment is executed later since
its behavior is dependent upon using the correct
Java VM.

If the Java VM used to bootstrap narwhal is not
in the PATH, append JAVA_HOME to narwhal.conf
so that narwhal can use this value.