Adding metadata for ivars.
This allows us to see what methods are associated with which ivars. I am going to use this with OJTest to auto-test setters and getters. While this does add overhead to Cappuccino, in general, the overhead should be as minimal as possible. The reason is that the class definitions exist in the global space and are only referenced, not created.

There are other possible uses revolving around metaprogramming. For example, you could use the ivar metadata to autogenerate relationships in model a la ActiveRecord.