The sender for invisible toolbar items was the menu item.
This fixes that the action sent to the target for a particular toolbar item would instead be the ephemeral CPMenuItem for invisible toolbar items. This prevents the target from determining the itemIdentifier which may be used in cases where all the toolbar items use the same target/action and rely on the sender to identify the toolbar item.