Fix Issue 1755: CPCollectionView drag update method
Without this fix, attempting to drag and drop an item into an empty CPCollectionView will fire `[_updateDragAndDropStateWithDraggingInfo:newDragOperation:newDropIndex:newDropOperation]`, which will first complain about a nil _dropView, and then will attempt to get the size of the previous item in the collection, which will be nil.

This commit fixes both errors by initializing the _dropView if it is not already initialized, and checking to see the length of the item count before attempting to determine the size of the previous item.